Create compendium helper
